requirements.txt
aiosqlite==0.18.0
flake8==6.0.0
isort==5.12.0
python-dotenv==1.0.0
python-telegram-bot==20.1